reafactor:platform related changes
