+++ pkg/platform/yaml/mockStore/util.go
@@ -1,98 +1,27 @@
-package yaml
+package mockstore
 
 import (
 	"errors"
-	"reflect"
-
 	"strings"
 
-	"go.keploy.io/server/v2/pkg"
 	"go.keploy.io/server/v2/pkg/models"
-	"go.keploy.io/server/v2/pkg/platform/yaml/spec"
+	"go.keploy.io/server/v2/pkg/platform/yaml"
+	"go.keploy.io/server/v2/utils"
 	"go.mongodb.org/mongo-driver/x/mongo/driver/wiremessage"
 	"go.uber.org/zap"
-	yamlLib "gopkg.in/yaml.v3"
 )
 
-// NetworkTrafficDoc stores the request-response data of a network call (ingress or egress)
-type NetworkTrafficDoc struct {
-	Version models.Version `json:"version" yaml:"version"`
-	Kind    models.Kind    `json:"kind" yaml:"kind"`
-	Name    string         `json:"name" yaml:"name"`
-	Spec    yamlLib.Node   `json:"spec" yaml:"spec"`
-	Curl    string         `json:"curl" yaml:"curl,omitempty"`
-}
-
-func (nd *NetworkTrafficDoc) GetKind() string {
-	return string(nd.Kind)
-}
-
-func EncodeTestcase(tc models.TestCase, logger *zap.Logger) (*NetworkTrafficDoc, error) {
-
-	header := pkg.ToHttpHeader(tc.HttpReq.Header)
-	curl := pkg.MakeCurlCommand(string(tc.HttpReq.Method), tc.HttpReq.URL, pkg.ToYamlHttpHeader(header), tc.HttpReq.Body)
-	doc := &NetworkTrafficDoc{
-		Version: tc.Version,
-		Kind:    tc.Kind,
-		Name:    tc.Name,
-		Curl:    curl,
-	}
-	// find noisy fields
-	m, err := FlattenHttpResponse(pkg.ToHttpHeader(tc.HttpResp.Header), tc.HttpResp.Body)
-	if err != nil {
-		msg := "error in flattening http response"
-		logger.Error(msg, zap.Error(err))
-	}
-	noise := tc.Noise
-
-	noiseFieldsFound := FindNoisyFields(m, func(k string, vals []string) bool {
-		// check if k is date
-		for _, v := range vals {
-			if pkg.IsTime(v) {
-				return true
-			}
-		}
-
-		// maybe we need to concatenate the values
-		return pkg.IsTime(strings.Join(vals, ", "))
-	})
-
-	for _, v := range noiseFieldsFound {
-		noise[v] = []string{}
-	}
-
-	switch tc.Kind {
-	case models.HTTP:
-		err := doc.Spec.Encode(spec.HttpSpec{
-			Request:  tc.HttpReq,
-			Response: tc.HttpResp,
-			Created:  tc.Created,
-			Assertions: map[string]interface{}{
-				"noise": noise,
-			},
-		})
-		if err != nil {
-			logger.Error("failed to encode testcase into a yaml doc", zap.Error(err))
-			return nil, err
-		}
-	default:
-		logger.Error("failed to marshal the testcase into yaml due to invalid kind of testcase")
-		return nil, errors.New("type of testcases is invalid")
-	}
-	return doc, nil
-}
-
-func EncodeMock(mock *models.Mock, logger *zap.Logger) (*NetworkTrafficDoc, error) {
-	yamlDoc := NetworkTrafficDoc{
+func EncodeMock(mock *models.Mock, logger *zap.Logger) (*yaml.NetworkTrafficDoc, error) {
+	yamlDoc := yaml.NetworkTrafficDoc{
 		Version: mock.Version,
 		Kind:    mock.Kind,
 		Name:    mock.Name,
 	}
 	switch mock.Kind {
 	case models.Mongo:
-		requests := []spec.RequestYaml{}
+		requests := []models.RequestYaml{}
 		for _, v := range mock.Spec.MongoRequests {
-			req := spec.RequestYaml{
+			req := models.RequestYaml{
 				Header:    v.Header,
 				ReadDelay: v.ReadDelay,
 			}
@@ -103,9 +32,9 @@ func EncodeMock(mock *models.Mock, logger *zap.Logger) (*NetworkTrafficDoc, erro
 			}
 			requests = append(requests, req)
 		}
-		responses := []spec.ResponseYaml{}
+		responses := []models.ResponseYaml{}
 		for _, v := range mock.Spec.MongoResponses {
-			resp := spec.ResponseYaml{
+			resp := models.ResponseYaml{
 				Header:    v.Header,
 				ReadDelay: v.ReadDelay,
 			}
@@ -116,7 +45,7 @@ func EncodeMock(mock *models.Mock, logger *zap.Logger) (*NetworkTrafficDoc, erro
 			}
 			responses = append(responses, resp)
 		}
-		mongoSpec := spec.MongoSpec{
+		mongoSpec := models.MongoSpec{
 			Metadata:         mock.Spec.Metadata,
 			Requests:         requests,
 			Response:         responses,
@@ -132,7 +61,7 @@ func EncodeMock(mock *models.Mock, logger *zap.Logger) (*NetworkTrafficDoc, erro
 		}
 
 	case models.HTTP:
-		httpSpec := spec.HttpSpec{
+		httpSpec := models.HttpSchema{
 			Metadata:         mock.Spec.Metadata,
 			Request:          *mock.Spec.HttpReq,
 			Response:         *mock.Spec.HttpResp,
@@ -146,7 +75,7 @@ func EncodeMock(mock *models.Mock, logger *zap.Logger) (*NetworkTrafficDoc, erro
 			return nil, err
 		}
 	case models.GENERIC:
-		genericSpec := spec.GenericSpec{
+		genericSpec := models.GenericSchema{
 			Metadata:         mock.Spec.Metadata,
 			GenericRequests:  mock.Spec.GenericRequests,
 			GenericResponses: mock.Spec.GenericResponses,
@@ -160,7 +89,7 @@ func EncodeMock(mock *models.Mock, logger *zap.Logger) (*NetworkTrafficDoc, erro
 		}
 	case models.Postgres:
 
-		postgresSpec := spec.PostgresSpec{
+		postgresSpec := models.PostgresSpec{
 			Metadata:          mock.Spec.Metadata,
 			PostgresRequests:  mock.Spec.PostgresRequests,
 			PostgresResponses: mock.Spec.PostgresResponses,
@@ -174,7 +103,7 @@ func EncodeMock(mock *models.Mock, logger *zap.Logger) (*NetworkTrafficDoc, erro
 			return nil, err
 		}
 	case models.GRPC_EXPORT:
-		gRPCSpec := spec.GrpcSpec{
+		gRPCSpec := models.GrpcSpec{
 			GrpcReq:          *mock.Spec.GRPCReq,
 			GrpcResp:         *mock.Spec.GRPCResp,
 			ReqTimestampMock: mock.Spec.ReqTimestampMock,
@@ -182,39 +111,39 @@ func EncodeMock(mock *models.Mock, logger *zap.Logger) (*NetworkTrafficDoc, erro
 		}
 		err := yamlDoc.Spec.Encode(gRPCSpec)
 		if err != nil {
-			logger.Error(Emoji+"failed to marshal gRPC of external call into yaml", zap.Error(err))
+			logger.Error(utils.Emoji+"failed to marshal gRPC of external call into yaml", zap.Error(err))
 			return nil, err
 		}
 	case models.SQL:
-		requests := []spec.MysqlRequestYaml{}
+		requests := []models.MysqlRequestYaml{}
 		for _, v := range mock.Spec.MySqlRequests {
 
-			req := spec.MysqlRequestYaml{
+			req := models.MysqlRequestYaml{
 				Header:    v.Header,
 				ReadDelay: v.ReadDelay,
 			}
 			err := req.Message.Encode(v.Message)
 			if err != nil {
-				logger.Error(Emoji+"failed to encode mongo request wiremessage into yaml", zap.Error(err))
+				logger.Error(utils.Emoji+"failed to encode mongo request wiremessage into yaml", zap.Error(err))
 				return nil, err
 			}
 			requests = append(requests, req)
 		}
-		responses := []spec.MysqlResponseYaml{}
+		responses := []models.MysqlResponseYaml{}
 		for _, v := range mock.Spec.MySqlResponses {
-			resp := spec.MysqlResponseYaml{
+			resp := models.MysqlResponseYaml{
 				Header:    v.Header,
 				ReadDelay: v.ReadDelay,
 			}
 			err := resp.Message.Encode(v.Message)
 			if err != nil {
-				logger.Error(Emoji+"failed to encode mongo request wiremessage into yaml", zap.Error(err))
+				logger.Error(utils.Emoji+"failed to encode mongo request wiremessage into yaml", zap.Error(err))
 				return nil, err
 			}
 			responses = append(responses, resp)
 		}
 
-		sqlSpec := spec.MySQLSpec{
+		sqlSpec := models.MySQLSpec{
 			Metadata:  mock.Spec.Metadata,
 			Requests:  requests,
 			Response:  responses,
@@ -222,7 +151,7 @@ func EncodeMock(mock *models.Mock, logger *zap.Logger) (*NetworkTrafficDoc, erro
 		}
 		err := yamlDoc.Spec.Encode(sqlSpec)
 		if err != nil {
-			logger.Error(Emoji+"failed to marshal the SQL input-output as yaml", zap.Error(err))
+			logger.Error(utils.Emoji+"failed to marshal the SQL input-output as yaml", zap.Error(err))
 			return nil, err
 		}
 	default:
@@ -233,56 +162,7 @@ func EncodeMock(mock *models.Mock, logger *zap.Logger) (*NetworkTrafficDoc, erro
 	return &yamlDoc, nil
 }
 
-func Decode(yamlTestcase *NetworkTrafficDoc, logger *zap.Logger) (*models.TestCase, error) {
-	tc := models.TestCase{
-		Version: yamlTestcase.Version,
-		Kind:    yamlTestcase.Kind,
-		Name:    yamlTestcase.Name,
-		Curl:    yamlTestcase.Curl,
-	}
-	switch tc.Kind {
-	case models.HTTP:
-		httpSpec := spec.HttpSpec{}
-		err := yamlTestcase.Spec.Decode(&httpSpec)
-		if err != nil {
-			logger.Error("failed to unmarshal a yaml doc into the http testcase", zap.Error(err))
-			return nil, err
-		}
-		tc.Created = httpSpec.Created
-		tc.HttpReq = httpSpec.Request
-		tc.HttpResp = httpSpec.Response
-		tc.Noise = map[string][]string{}
-		switch reflect.ValueOf(httpSpec.Assertions["noise"]).Kind() {
-		case reflect.Map:
-			for k, v := range httpSpec.Assertions["noise"].(map[string]interface{}) {
-				tc.Noise[k] = []string{}
-				for _, val := range v.([]interface{}) {
-					tc.Noise[k] = append(tc.Noise[k], val.(string))
-				}
-			}
-		case reflect.Slice:
-			for _, v := range httpSpec.Assertions["noise"].([]interface{}) {
-				tc.Noise[v.(string)] = []string{}
-			}
-		}
-	// unmarshal its mocks from yaml docs to go struct
-	case models.GRPC_EXPORT:
-		grpcSpec := spec.GrpcSpec{}
-		err := yamlTestcase.Spec.Decode(&grpcSpec)
-		if err != nil {
-			logger.Error(Emoji+"failed to unmarshal a yaml doc into the gRPC testcase", zap.Error(err))
-			return nil, err
-		}
-		tc.GrpcReq = grpcSpec.GrpcReq
-		tc.GrpcResp = grpcSpec.GrpcResp
-	default:
-		logger.Error("failed to unmarshal yaml doc of unknown type", zap.Any("type of yaml doc", tc.Kind))
-		return nil, errors.New("yaml doc of unknown type")
-	}
-	return &tc, nil
-}
-
-func decodeMocks(yamlMocks []*NetworkTrafficDoc, logger *zap.Logger) ([]*models.Mock, error) {
+func decodeMocks(yamlMocks []*yaml.NetworkTrafficDoc, logger *zap.Logger) ([]*models.Mock, error) {
 	mocks := []*models.Mock{}
 
 	for _, m := range yamlMocks {
@@ -298,7 +178,7 @@ func decodeMocks(yamlMocks []*NetworkTrafficDoc, logger *zap.Logger) ([]*models.
 		}
 		switch m.Kind {
 		case models.HTTP:
-			httpSpec := spec.HttpSpec{}
+			httpSpec := models.HttpSchema{}
 			err := m.Spec.Decode(&httpSpec)
 			if err != nil {
 				logger.Error("failed to unmarshal a yaml doc into http mock", zap.Error(err), zap.Any("mock name", m.Name))
@@ -314,7 +194,7 @@ func decodeMocks(yamlMocks []*NetworkTrafficDoc, logger *zap.Logger) ([]*models.
 				ResTimestampMock: httpSpec.ResTimestampMock,
 			}
 		case models.Mongo:
-			mongoSpec := spec.MongoSpec{}
+			mongoSpec := models.MongoSpec{}
 			err := m.Spec.Decode(&mongoSpec)
 			if err != nil {
 				logger.Error("failed to unmarshal a yaml doc into mongo mock", zap.Error(err), zap.Any("mock name", m.Name))
@@ -327,10 +207,10 @@ func decodeMocks(yamlMocks []*NetworkTrafficDoc, logger *zap.Logger) ([]*models.
 			}
 			mock.Spec = *mockSpec
 		case models.GRPC_EXPORT:
-			grpcSpec := spec.GrpcSpec{}
+			grpcSpec := models.GrpcSpec{}
 			err := m.Spec.Decode(&grpcSpec)
 			if err != nil {
-				logger.Error(Emoji+"failed to unmarshal a yaml doc into http mock", zap.Error(err), zap.Any("mock name", m.Name))
+				logger.Error(utils.Emoji+"failed to unmarshal a yaml doc into http mock", zap.Error(err), zap.Any("mock name", m.Name))
 				return nil, err
 			}
 			mock.Spec = models.MockSpec{
@@ -340,7 +220,7 @@ func decodeMocks(yamlMocks []*NetworkTrafficDoc, logger *zap.Logger) ([]*models.
 				ResTimestampMock: grpcSpec.ResTimestampMock,
 			}
 		case models.GENERIC:
-			genericSpec := spec.GenericSpec{}
+			genericSpec := models.GenericSchema{}
 			err := m.Spec.Decode(&genericSpec)
 			if err != nil {
 				logger.Error("failed to unmarshal a yaml doc into generic mock", zap.Error(err), zap.Any("mock name", m.Name))
@@ -356,7 +236,7 @@ func decodeMocks(yamlMocks []*NetworkTrafficDoc, logger *zap.Logger) ([]*models.
 
 		case models.Postgres:
 
-			PostSpec := spec.PostgresSpec{}
+			PostSpec := models.PostgresSpec{}
 			err := m.Spec.Decode(&PostSpec)
 
 			if err != nil {
@@ -372,10 +252,10 @@ func decodeMocks(yamlMocks []*NetworkTrafficDoc, logger *zap.Logger) ([]*models.
 				ResTimestampMock:  PostSpec.ResTimestampMock,
 			}
 		case models.SQL:
-			mysqlSpec := spec.MySQLSpec{}
+			mysqlSpec := models.MySQLSpec{}
 			err := m.Spec.Decode(&mysqlSpec)
 			if err != nil {
-				logger.Error(Emoji+"failed to unmarshal a yaml doc into mongo mock", zap.Error(err), zap.Any("mock name", m.Name))
+				logger.Error(utils.Emoji+"failed to unmarshal a yaml doc into mongo mock", zap.Error(err), zap.Any("mock name", m.Name))
 				return nil, err
 			}
 
@@ -393,7 +273,8 @@ func decodeMocks(yamlMocks []*NetworkTrafficDoc, logger *zap.Logger) ([]*models.
 
 	return mocks, nil
 }
-func decodeMySqlMessage(yamlSpec *spec.MySQLSpec, logger *zap.Logger) (*models.MockSpec, error) {
+
+func decodeMySqlMessage(yamlSpec *models.MySQLSpec, logger *zap.Logger) (*models.MockSpec, error) {
 	mockSpec := models.MockSpec{
 		Metadata: yamlSpec.Metadata,
 		Created:  yamlSpec.CreatedAt,
@@ -409,7 +290,7 @@ func decodeMySqlMessage(yamlSpec *spec.MySQLSpec, logger *zap.Logger) (*models.M
 			requestMessage := &models.MySQLHandshakeResponse{}
 			err := v.Message.Decode(requestMessage)
 			if err != nil {
-				logger.Error(Emoji+"failed to unmarshal yml document into MySQLHandshakeResponse ", zap.Error(err))
+				logger.Error(utils.Emoji+"failed to unmarshal yml document into MySQLHandshakeResponse ", zap.Error(err))
 				return nil, err
 			}
 			req.Message = requestMessage
@@ -417,7 +298,7 @@ func decodeMySqlMessage(yamlSpec *spec.MySQLSpec, logger *zap.Logger) (*models.M
 			requestMessage := &models.MySQLQueryPacket{}
 			err := v.Message.Decode(requestMessage)
 			if err != nil {
-				logger.Error(Emoji+"failed to unmarshal yml document into MySQLQueryPacket", zap.Error(err))
+				logger.Error(utils.Emoji+"failed to unmarshal yml document into MySQLQueryPacket", zap.Error(err))
 				return nil, err
 			}
 			req.Message = requestMessage
@@ -425,7 +306,7 @@ func decodeMySqlMessage(yamlSpec *spec.MySQLSpec, logger *zap.Logger) (*models.M
 			requestMessage := &models.MySQLComStmtPreparePacket{}
 			err := v.Message.Decode(requestMessage)
 			if err != nil {
-				logger.Error(Emoji+"failed to unmarshal yml document into MySQLComStmtPreparePacket", zap.Error(err))
+				logger.Error(utils.Emoji+"failed to unmarshal yml document into MySQLComStmtPreparePacket", zap.Error(err))
 				return nil, err
 			}
 			req.Message = requestMessage
@@ -433,7 +314,7 @@ func decodeMySqlMessage(yamlSpec *spec.MySQLSpec, logger *zap.Logger) (*models.M
 			requestMessage := &models.MySQLComStmtExecute{}
 			err := v.Message.Decode(requestMessage)
 			if err != nil {
-				logger.Error(Emoji+"failed to unmarshal yml document into MySQLComStmtExecute", zap.Error(err))
+				logger.Error(utils.Emoji+"failed to unmarshal yml document into MySQLComStmtExecute", zap.Error(err))
 				return nil, err
 			}
 			req.Message = requestMessage
@@ -441,7 +322,7 @@ func decodeMySqlMessage(yamlSpec *spec.MySQLSpec, logger *zap.Logger) (*models.M
 			requestMessage := &models.MySQLCOM_STMT_SEND_LONG_DATA{}
 			err := v.Message.Decode(requestMessage)
 			if err != nil {
-				logger.Error(Emoji+"failed to unmarshal yml document into MySQLCOM_STMT_SEND_LONG_DATA", zap.Error(err))
+				logger.Error(utils.Emoji+"failed to unmarshal yml document into MySQLCOM_STMT_SEND_LONG_DATA", zap.Error(err))
 				return nil, err
 			}
 			req.Message = requestMessage
@@ -449,7 +330,7 @@ func decodeMySqlMessage(yamlSpec *spec.MySQLSpec, logger *zap.Logger) (*models.M
 			requestMessage := &models.MySQLCOM_STMT_RESET{}
 			err := v.Message.Decode(requestMessage)
 			if err != nil {
-				logger.Error(Emoji+"failed to unmarshal yml document into MySQLCOM_STMT_RESET", zap.Error(err))
+				logger.Error(utils.Emoji+"failed to unmarshal yml document into MySQLCOM_STMT_RESET", zap.Error(err))
 				return nil, err
 			}
 			req.Message = requestMessage
@@ -457,7 +338,7 @@ func decodeMySqlMessage(yamlSpec *spec.MySQLSpec, logger *zap.Logger) (*models.M
 			requestMessage := &models.MySQLComStmtFetchPacket{}
 			err := v.Message.Decode(requestMessage)
 			if err != nil {
-				logger.Error(Emoji+"failed to unmarshal yml document into MySQLComStmtFetchPacket", zap.Error(err))
+				logger.Error(utils.Emoji+"failed to unmarshal yml document into MySQLComStmtFetchPacket", zap.Error(err))
 				return nil, err
 			}
 			req.Message = requestMessage
@@ -465,7 +346,7 @@ func decodeMySqlMessage(yamlSpec *spec.MySQLSpec, logger *zap.Logger) (*models.M
 			requestMessage := &models.MySQLComStmtClosePacket{}
 			err := v.Message.Decode(requestMessage)
 			if err != nil {
-				logger.Error(Emoji+"failed to unmarshal yml document into MySQLComStmtClosePacket", zap.Error(err))
+				logger.Error(utils.Emoji+"failed to unmarshal yml document into MySQLComStmtClosePacket", zap.Error(err))
 				return nil, err
 			}
 			req.Message = requestMessage
@@ -473,7 +354,7 @@ func decodeMySqlMessage(yamlSpec *spec.MySQLSpec, logger *zap.Logger) (*models.M
 			requestMessage := &models.AuthSwitchRequestPacket{}
 			err := v.Message.Decode(requestMessage)
 			if err != nil {
-				logger.Error(Emoji+"failed to unmarshal yml document into MySQLComStmtClosePacket", zap.Error(err))
+				logger.Error(utils.Emoji+"failed to unmarshal yml document into MySQLComStmtClosePacket", zap.Error(err))
 				return nil, err
 			}
 			req.Message = requestMessage
@@ -481,7 +362,7 @@ func decodeMySqlMessage(yamlSpec *spec.MySQLSpec, logger *zap.Logger) (*models.M
 			requestMessage := &models.MySQLComChangeUserPacket{}
 			err := v.Message.Decode(requestMessage)
 			if err != nil {
-				logger.Error(Emoji+"failed to unmarshal yml document into MySQLComChangeUserPacket", zap.Error(err))
+				logger.Error(utils.Emoji+"failed to unmarshal yml document into MySQLComChangeUserPacket", zap.Error(err))
 				return nil, err
 			}
 			req.Message = requestMessage
@@ -502,7 +383,7 @@ func decodeMySqlMessage(yamlSpec *spec.MySQLSpec, logger *zap.Logger) (*models.M
 			responseMessage := &models.MySQLHandshakeResponseOk{}
 			err := v.Message.Decode(responseMessage)
 			if err != nil {
-				logger.Error(Emoji+"failed to unmarshal yml document into MySQLHandshakeResponseOk ", zap.Error(err))
+				logger.Error(utils.Emoji+"failed to unmarshal yml document into MySQLHandshakeResponseOk ", zap.Error(err))
 				return nil, err
 			}
 			resp.Message = responseMessage
@@ -510,7 +391,7 @@ func decodeMySqlMessage(yamlSpec *spec.MySQLSpec, logger *zap.Logger) (*models.M
 			responseMessage := &models.MySQLHandshakeV10Packet{}
 			err := v.Message.Decode(responseMessage)
 			if err != nil {
-				logger.Error(Emoji+"failed to unmarshal yml document into MySQLHandshakeV10Packet", zap.Error(err))
+				logger.Error(utils.Emoji+"failed to unmarshal yml document into MySQLHandshakeV10Packet", zap.Error(err))
 				return nil, err
 			}
 			resp.Message = responseMessage
@@ -518,7 +399,7 @@ func decodeMySqlMessage(yamlSpec *spec.MySQLSpec, logger *zap.Logger) (*models.M
 			responseMessage := &models.MySQLOKPacket{}
 			err := v.Message.Decode(responseMessage)
 			if err != nil {
-				logger.Error(Emoji+"failed to unmarshal yml document into MySQLOKPacket ", zap.Error(err))
+				logger.Error(utils.Emoji+"failed to unmarshal yml document into MySQLOKPacket ", zap.Error(err))
 				return nil, err
 			}
 			resp.Message = responseMessage
@@ -526,7 +407,7 @@ func decodeMySqlMessage(yamlSpec *spec.MySQLSpec, logger *zap.Logger) (*models.M
 			responseMessage := &models.MySQLStmtPrepareOk{}
 			err := v.Message.Decode(responseMessage)
 			if err != nil {
-				logger.Error(Emoji+"failed to unmarshal yml document into MySQLStmtPrepareOk ", zap.Error(err))
+				logger.Error(utils.Emoji+"failed to unmarshal yml document into MySQLStmtPrepareOk ", zap.Error(err))
 				return nil, err
 			}
 			resp.Message = responseMessage
@@ -534,7 +415,7 @@ func decodeMySqlMessage(yamlSpec *spec.MySQLSpec, logger *zap.Logger) (*models.M
 			responseMessage := &models.MySQLResultSet{}
 			err := v.Message.Decode(responseMessage)
 			if err != nil {
-				logger.Error(Emoji+"failed to unmarshal yml document into MySQLResultSet ", zap.Error(err))
+				logger.Error(utils.Emoji+"failed to unmarshal yml document into MySQLResultSet ", zap.Error(err))
 				return nil, err
 			}
 			resp.Message = responseMessage
@@ -542,7 +423,7 @@ func decodeMySqlMessage(yamlSpec *spec.MySQLSpec, logger *zap.Logger) (*models.M
 			responseMessage := &models.AuthSwitchRequestPacket{}
 			err := v.Message.Decode(responseMessage)
 			if err != nil {
-				logger.Error(Emoji+"failed to unmarshal yml document into MySQLResultSet ", zap.Error(err))
+				logger.Error(utils.Emoji+"failed to unmarshal yml document into MySQLResultSet ", zap.Error(err))
 				return nil, err
 			}
 			resp.Message = responseMessage
@@ -550,7 +431,7 @@ func decodeMySqlMessage(yamlSpec *spec.MySQLSpec, logger *zap.Logger) (*models.M
 			responseMessage := &models.MySQLERRPacket{}
 			err := v.Message.Decode(responseMessage)
 			if err != nil {
-				logger.Error(Emoji+"failed to unmarshal yml document into MySQLERRPacket ", zap.Error(err))
+				logger.Error(utils.Emoji+"failed to unmarshal yml document into MySQLERRPacket ", zap.Error(err))
 				return nil, err
 			}
 			resp.Message = responseMessage
@@ -561,7 +442,7 @@ func decodeMySqlMessage(yamlSpec *spec.MySQLSpec, logger *zap.Logger) (*models.M
 	return &mockSpec, nil
 
 }
-func decodeMongoMessage(yamlSpec *spec.MongoSpec, logger *zap.Logger) (*models.MockSpec, error) {
+func decodeMongoMessage(yamlSpec *models.MongoSpec, logger *zap.Logger) (*models.MockSpec, error) {
 	mockSpec := models.MockSpec{
 		Metadata:         yamlSpec.Metadata,
 		Created:          yamlSpec.CreatedAt,
