--- pkg/platform/yaml/spec/README.md
@@ -1,5 +1,0 @@
-# Spec Package Documentation
-
-This package contains Go structs for all the different types of API input-output
-messages. Designed as an independent module, its primary use is 
-within the `platform/yaml` package.
