+++ utils/utils.go
@@ -439,3 +439,19 @@ func GetUniqueReportDir(testReportPath, subDirPrefix string) (string, error) {
 	return newTestReportPath, nil
 }
 
+func FetchHomeDirectory(isNewConfigPath bool) string {
+	var configFolder = "/.keploy-config"
+
+	if isNewConfigPath {
+		configFolder = "/.keploy"
+	}
+	if runtime.GOOS == "windows" {
+		home := os.Getenv("HOMEDRIVE") + os.Getenv("HOMEPATH")
+		if home == "" {
+			home = os.Getenv("USERPROFILE")
+		}
+		return home + configFolder
+	}
+
+	return os.Getenv("HOME") + configFolder
+}
